refactored the backend for later merge to temp
