+++ BackEnd/src/main/java/com/example/backend/controller/RecordController.java
@@ -1,5 +1,7 @@
-package com.example.backend.stuff;
+package com.example.backend.controller;
 
+import com.example.backend.model.Record;
+import com.example.backend.repository.RecordRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
