+++ BackEnd/src/main/java/com/example/backend/model/Doctor.java
@@ -1,4 +1,4 @@
-package com.example.backend.stuff;
+package com.example.backend.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
