--- BackEnd/src/main/java/com/example/backend/stuff/UserJPARepository.java
@@ -1,7 +1,0 @@
-package com.example.backend.stuff;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserJPARepository extends JpaRepository<User,Integer> {
-    User findByEmailAndPassword(String email, String password);
-}
