+++ BackEnd/src/main/resources/application.properties
@@ -1,10 +1,10 @@
 #spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-#spring.datasource.username=root
-#spring.datasource.password=RNXNRngu7898
-#spring.datasource.url=jdbc:mysql://127.0.0.1:3306/sept
-spring.datasource.username=admin_jason
-spring.datasource.password=RMIT!1234
-spring.datasource.url=jdbc:mysql://database-1.cavyvvu6hnkp.us-east-1.rds.amazonaws.com/database-1?characterEncoding=utf8
+spring.datasource.username=root
+spring.datasource.password=RNXNRngu7898
+spring.datasource.url=jdbc:mysql://127.0.0.1:3306/sept
+#spring.datasource.username=admin_jason
+#spring.datasource.password=RMIT!1234
+#spring.datasource.url=jdbc:mysql://database-1.cavyvvu6hnkp.us-east-1.rds.amazonaws.com/database-1?characterEncoding=utf8
